Remove all dynamic content filtering.
Selectively including content was becoming overly-complex, especially
since determining inclusion rules was purely empirical. As the output
of this program is intended for a diff tool, simply including everything
is preferable to incorrect omission as the diff tool won't flag things
that didn't change, even if they are irrelevant. Determining what
parameters are applicable is the domain of SISTEMA; a task this program
expressly does not attempt to duplicate.